PHP 8.1: fix deprecation warnings
Came across these:

```
    [phpab] PHP Deprecated:  Return type of TheSeer\DirectoryScanner\IncludeExcludeFilterIterator::accept() should either be compatible with FilterIterator::accept(): bool, or the #[ReturnTypeWillChange] attribute should be used to temporarily suppress the notice in phar:///home/runner/work/phpunit/phpunit/tools/phpab/vendor/theseer/directoryscanner/src/includeexcludefilter.php on line 91
    [phpab] PHP Deprecated:  Return type of TheSeer\DirectoryScanner\FilesOnlyFilterIterator::accept() should either be compatible with FilterIterator::accept(): bool, or the #[ReturnTypeWillChange] attribute should be used to temporarily suppress the notice in phar:///home/runner/work/phpunit/phpunit/tools/phpab/vendor/theseer/directoryscanner/src/filesonlyfilter.php on line 55
```

These type of deprecation notices relate to the [Return types for internal methods RFC](https://wiki.php.net/rfc/internal_method_return_types) in PHP 8.1.

Basically, as of PHP 8.1, these methods in classes which implement PHP native interfaces are expected to have a return type declared.
The return type can be the same as used in PHP itself or a more specific type. This complies with the Liskov principle of covariance, which allows the return type of a child overloaded method to be more specific than that of the parent.

As this package has a minimum PHP requirement of PHP 5.3 based on the `composer.json` file, the return type can not be added, so I've added the attribute instead.

While attributes are a PHP 8.0 feature only, due to the syntax choice for `#[]`, they will ignored in PHP < 8 and can be safely added.